+++ server/runtime_lua_nakama.go
@@ -58,7 +58,7 @@ type RuntimeLuaNakamaModule struct {
 	leaderboardCache     LeaderboardCache
 	rankCache            LeaderboardRankCache
 	leaderboardScheduler LeaderboardScheduler
-	sessionRegistry      *SessionRegistry
+	sessionRegistry      SessionRegistry
 	matchRegistry        MatchRegistry
 	tracker              Tracker
 	router               MessageRouter
@@ -72,7 +72,7 @@ type RuntimeLuaNakamaModule struct {
 	matchCreateFn RuntimeMatchCreateFunction
 }
 
-func NewRuntimeLuaNakamaModule(logger *zap.Logger, db *sql.DB, jsonpbUnmarshaler *jsonpb.Unmarshaler, config Config, socialClient *social.Client, leaderboardCache LeaderboardCache, rankCache LeaderboardRankCache, leaderboardScheduler LeaderboardScheduler, sessionRegistry *SessionRegistry, matchRegistry MatchRegistry, tracker Tracker, router MessageRouter, once *sync.Once, localCache *RuntimeLuaLocalCache, matchCreateFn RuntimeMatchCreateFunction, registerCallbackFn func(RuntimeExecutionMode, string, *lua.LFunction), announceCallbackFn func(RuntimeExecutionMode, string)) *RuntimeLuaNakamaModule {
+func NewRuntimeLuaNakamaModule(logger *zap.Logger, db *sql.DB, jsonpbUnmarshaler *jsonpb.Unmarshaler, config Config, socialClient *social.Client, leaderboardCache LeaderboardCache, rankCache LeaderboardRankCache, leaderboardScheduler LeaderboardScheduler, sessionRegistry SessionRegistry, matchRegistry MatchRegistry, tracker Tracker, router MessageRouter, once *sync.Once, localCache *RuntimeLuaLocalCache, matchCreateFn RuntimeMatchCreateFunction, registerCallbackFn func(RuntimeExecutionMode, string, *lua.LFunction), announceCallbackFn func(RuntimeExecutionMode, string)) *RuntimeLuaNakamaModule {
 	return &RuntimeLuaNakamaModule{
 		logger:               logger,
 		db:                   db,
@@ -166,6 +166,7 @@ func (n *RuntimeLuaNakamaModule) Loader(l *lua.LState) int {
 		"stream_close":                n.streamClose,
 		"stream_send":                 n.streamSend,
 		"stream_send_raw":             n.streamSendRaw,
+		"session_disconnect":          n.sessionDisconnect,
 		"match_create":                n.matchCreate,
 		"match_list":                  n.matchList,
 		"notification_send":           n.notificationSend,
@@ -2064,10 +2065,13 @@ func (n *RuntimeLuaNakamaModule) streamUserJoin(l *lua.LState) int {
 		return 0
 	}
 
+	// Parse Node.
+	node := l.OptString(3, n.node)
+
 	// Parse input stream identifier.
-	streamTable := l.CheckTable(3)
+	streamTable := l.CheckTable(4)
 	if streamTable == nil {
-		l.ArgError(3, "expects a valid stream")
+		l.ArgError(4, "expects a valid stream")
 		return 0
 	}
 	stream := PresenceStream{}
@@ -2081,40 +2085,40 @@ func (n *RuntimeLuaNakamaModule) streamUserJoin(l *lua.LState) int {
 		case "mode":
 			if v.Type() != lua.LTNumber {
 				conversionError = true
-				l.ArgError(3, "stream mode must be a number")
+				l.ArgError(4, "stream mode must be a number")
 				return
 			}
 			stream.Mode = uint8(lua.LVAsNumber(v))
 		case "subject":
 			if v.Type() != lua.LTString {
 				conversionError = true
-				l.ArgError(3, "stream subject must be a string")
+				l.ArgError(4, "stream subject must be a string")
 				return
 			}
 			sid, err := uuid.FromString(v.String())
 			if err != nil {
 				conversionError = true
-				l.ArgError(3, "stream subject must be a valid identifier")
+				l.ArgError(4, "stream subject must be a valid identifier")
 				return
 			}
 			stream.Subject = sid
 		case "subcontext":
 			if v.Type() != lua.LTString {
 				conversionError = true
-				l.ArgError(3, "stream subcontext must be a string")
+				l.ArgError(4, "stream subcontext must be a string")
 				return
 			}
 			sid, err := uuid.FromString(v.String())
 			if err != nil {
 				conversionError = true
-				l.ArgError(3, "stream subcontext must be a valid identifier")
+				l.ArgError(4, "stream subcontext must be a valid identifier")
 				return
 			}
 			stream.Subcontext = sid
 		case "label":
 			if v.Type() != lua.LTString {
 				conversionError = true
-				l.ArgError(3, "stream label must be a string")
+				l.ArgError(4, "stream label must be a string")
 				return
 			}
 			stream.Label = v.String()
@@ -2125,11 +2129,11 @@ func (n *RuntimeLuaNakamaModule) streamUserJoin(l *lua.LState) int {
 	}
 
 	// By default generate presence events.
-	hidden := l.OptBool(4, false)
+	hidden := l.OptBool(5, false)
 	// By default persistence is enabled, if the stream supports it.
-	persistence := l.OptBool(5, true)
+	persistence := l.OptBool(6, true)
 	// By default no status is set.
-	status := l.OptString(6, "")
+	status := l.OptString(7, "")
 
 	// Look up the session.
 	session := n.sessionRegistry.Get(sessionID)
@@ -2138,7 +2142,7 @@ func (n *RuntimeLuaNakamaModule) streamUserJoin(l *lua.LState) int {
 		return 0
 	}
 
-	success, newlyTracked := n.tracker.Track(sessionID, stream, userID, PresenceMeta{
+	success, newlyTracked := n.tracker.Track(sessionID, stream, userID, node, PresenceMeta{
 		Format:      session.Format(),
 		Hidden:      hidden,
 		Persistence: persistence,
@@ -2179,10 +2183,13 @@ func (n *RuntimeLuaNakamaModule) streamUserUpdate(l *lua.LState) int {
 		return 0
 	}
 
+	// Parse Node.
+	node := l.OptString(3, n.node)
+
 	// Parse input stream identifier.
-	streamTable := l.CheckTable(3)
+	streamTable := l.CheckTable(4)
 	if streamTable == nil {
-		l.ArgError(3, "expects a valid stream")
+		l.ArgError(4, "expects a valid stream")
 		return 0
 	}
 	stream := PresenceStream{}
@@ -2196,40 +2203,40 @@ func (n *RuntimeLuaNakamaModule) streamUserUpdate(l *lua.LState) int {
 		case "mode":
 			if v.Type() != lua.LTNumber {
 				conversionError = true
-				l.ArgError(3, "stream mode must be a number")
+				l.ArgError(4, "stream mode must be a number")
 				return
 			}
 			stream.Mode = uint8(lua.LVAsNumber(v))
 		case "subject":
 			if v.Type() != lua.LTString {
 				conversionError = true
-				l.ArgError(3, "stream subject must be a string")
+				l.ArgError(4, "stream subject must be a string")
 				return
 			}
 			sid, err := uuid.FromString(v.String())
 			if err != nil {
 				conversionError = true
-				l.ArgError(3, "stream subject must be a valid identifier")
+				l.ArgError(4, "stream subject must be a valid identifier")
 				return
 			}
 			stream.Subject = sid
 		case "subcontext":
 			if v.Type() != lua.LTString {
 				conversionError = true
-				l.ArgError(3, "stream subcontext must be a string")
+				l.ArgError(4, "stream subcontext must be a string")
 				return
 			}
 			sid, err := uuid.FromString(v.String())
 			if err != nil {
 				conversionError = true
-				l.ArgError(3, "stream subcontext must be a valid identifier")
+				l.ArgError(4, "stream subcontext must be a valid identifier")
 				return
 			}
 			stream.Subcontext = sid
 		case "label":
 			if v.Type() != lua.LTString {
 				conversionError = true
-				l.ArgError(3, "stream label must be a string")
+				l.ArgError(4, "stream label must be a string")
 				return
 			}
 			stream.Label = v.String()
@@ -2240,11 +2247,11 @@ func (n *RuntimeLuaNakamaModule) streamUserUpdate(l *lua.LState) int {
 	}
 
 	// By default generate presence events.
-	hidden := l.OptBool(4, false)
+	hidden := l.OptBool(5, false)
 	// By default persistence is enabled, if the stream supports it.
-	persistence := l.OptBool(5, true)
+	persistence := l.OptBool(6, true)
 	// By default no status is set.
-	status := l.OptString(6, "")
+	status := l.OptString(7, "")
 
 	// Look up the session.
 	session := n.sessionRegistry.Get(sessionID)
@@ -2253,7 +2260,7 @@ func (n *RuntimeLuaNakamaModule) streamUserUpdate(l *lua.LState) int {
 		return 0
 	}
 
-	if !n.tracker.Update(sessionID, stream, userID, PresenceMeta{
+	if !n.tracker.Update(sessionID, stream, userID, node, PresenceMeta{
 		Format:      session.Format(),
 		Hidden:      hidden,
 		Persistence: persistence,
@@ -2291,10 +2298,13 @@ func (n *RuntimeLuaNakamaModule) streamUserLeave(l *lua.LState) int {
 		return 0
 	}
 
+	// Parse Node.
+	node := l.OptString(3, n.node)
+
 	// Parse input stream identifier.
-	streamTable := l.CheckTable(3)
+	streamTable := l.CheckTable(4)
 	if streamTable == nil {
-		l.ArgError(3, "expects a valid stream")
+		l.ArgError(4, "expects a valid stream")
 		return 0
 	}
 	stream := PresenceStream{}
@@ -2308,40 +2318,40 @@ func (n *RuntimeLuaNakamaModule) streamUserLeave(l *lua.LState) int {
 		case "mode":
 			if v.Type() != lua.LTNumber {
 				conversionError = true
-				l.ArgError(3, "stream mode must be a number")
+				l.ArgError(4, "stream mode must be a number")
 				return
 			}
 			stream.Mode = uint8(lua.LVAsNumber(v))
 		case "subject":
 			if v.Type() != lua.LTString {
 				conversionError = true
-				l.ArgError(3, "stream subject must be a string")
+				l.ArgError(4, "stream subject must be a string")
 				return
 			}
 			sid, err := uuid.FromString(v.String())
 			if err != nil {
 				conversionError = true
-				l.ArgError(3, "stream subject must be a valid identifier")
+				l.ArgError(4, "stream subject must be a valid identifier")
 				return
 			}
 			stream.Subject = sid
 		case "subcontext":
 			if v.Type() != lua.LTString {
 				conversionError = true
-				l.ArgError(3, "stream subcontext must be a string")
+				l.ArgError(4, "stream subcontext must be a string")
 				return
 			}
 			sid, err := uuid.FromString(v.String())
 			if err != nil {
 				conversionError = true
-				l.ArgError(3, "stream subcontext must be a valid identifier")
+				l.ArgError(4, "stream subcontext must be a valid identifier")
 				return
 			}
 			stream.Subcontext = sid
 		case "label":
 			if v.Type() != lua.LTString {
 				conversionError = true
-				l.ArgError(3, "stream label must be a string")
+				l.ArgError(4, "stream label must be a string")
 				return
 			}
 			stream.Label = v.String()
@@ -2351,7 +2361,7 @@ func (n *RuntimeLuaNakamaModule) streamUserLeave(l *lua.LState) int {
 		return 0
 	}
 
-	n.tracker.Untrack(sessionID, stream, userID)
+	n.tracker.Untrack(sessionID, stream, userID, node)
 
 	return 0
 }
@@ -2553,6 +2563,68 @@ func (n *RuntimeLuaNakamaModule) streamSend(l *lua.LState) int {
 	// Grab payload to send, allow empty data.
 	data := l.CheckString(2)
 
+	presencesTable := l.OptTable(3, nil)
+	var presenceIDs []*PresenceID
+	if presencesTable != nil {
+		if ln := presencesTable.Len(); ln != 0 {
+			presenceIDs = make([]*PresenceID, 0, ln)
+			presencesTable.ForEach(func(k lua.LValue, v lua.LValue) {
+				if conversionError {
+					return
+				}
+
+				presenceTable, ok := v.(*lua.LTable)
+				if !ok {
+					conversionError = true
+					l.ArgError(3, "expects a valid set of presences")
+					return
+				}
+
+				presenceID := &PresenceID{}
+				presenceTable.ForEach(func(k lua.LValue, v lua.LValue) {
+					if conversionError {
+						return
+					}
+
+					switch k.String() {
+					case "session_id":
+						if v.Type() != lua.LTString {
+							conversionError = true
+							l.ArgError(3, "presence session id must be a string")
+							return
+						}
+						var err error
+						presenceID.SessionID, err = uuid.FromString(v.String())
+						if err != nil {
+							conversionError = true
+							l.ArgError(3, "presence session id must be a valid identifier")
+							return
+						}
+					case "node_id":
+						if v.Type() != lua.LTString {
+							conversionError = true
+							l.ArgError(3, "presence node id must be a string")
+							return
+						}
+						presenceID.Node = v.String()
+					}
+				})
+				if conversionError {
+					return
+				}
+
+				if presenceID.Node == "" {
+					presenceID.Node = n.node
+				}
+
+				presenceIDs = append(presenceIDs, presenceID)
+			})
+		}
+	}
+	if conversionError {
+		return 0
+	}
+
 	streamWire := &rtapi.Stream{
 		Mode:  int32(stream.Mode),
 		Label: stream.Label,
@@ -2568,7 +2640,14 @@ func (n *RuntimeLuaNakamaModule) streamSend(l *lua.LState) int {
 		// No sender.
 		Data: data,
 	}}}
-	n.router.SendToStream(n.logger, stream, msg)
+
+	if len(presenceIDs) == 0 {
+		// Sending to whole stream.
+		n.router.SendToStream(n.logger, stream, msg)
+	} else {
+		// Sending to a subset of stream users.
+		n.router.SendToPresenceIDs(n.logger, presenceIDs, true, stream.Mode, msg)
+	}
 
 	return 0
 }
@@ -2634,6 +2713,7 @@ func (n *RuntimeLuaNakamaModule) streamSendRaw(l *lua.LState) int {
 		return 0
 	}
 
+	// Parse the envelope.
 	envelopeMap := RuntimeLuaConvertLuaTable(l.CheckTable(2))
 	envelopeBytes, err := json.Marshal(envelopeMap)
 	if err != nil {
@@ -2641,25 +2721,103 @@ func (n *RuntimeLuaNakamaModule) streamSendRaw(l *lua.LState) int {
 		return 0
 	}
 
-	envelope := &rtapi.Envelope{}
-	err = n.jsonpbUnmarshaler.Unmarshal(bytes.NewReader(envelopeBytes), envelope)
-	if err != nil {
+	msg := &rtapi.Envelope{}
+	if err = n.jsonpbUnmarshaler.Unmarshal(bytes.NewReader(envelopeBytes), msg); err != nil {
 		l.ArgError(2, fmt.Sprintf("not a valid envlope: %s", err.Error()))
 		return 0
 	}
 
-	streamWire := &rtapi.Stream{
-		Mode:  int32(stream.Mode),
-		Label: stream.Label,
-	}
-	if stream.Subject != uuid.Nil {
-		streamWire.Subject = stream.Subject.String()
-	}
-	if stream.Subcontext != uuid.Nil {
-		streamWire.Subcontext = stream.Subcontext.String()
-	}
-	n.router.SendToStream(n.logger, stream, envelope)
+	// Validate subset of presences, if any.
+	presencesTable := l.OptTable(3, nil)
+	var presenceIDs []*PresenceID
+	if presencesTable != nil {
+		if ln := presencesTable.Len(); ln != 0 {
+			presenceIDs = make([]*PresenceID, 0, ln)
+			presencesTable.ForEach(func(k lua.LValue, v lua.LValue) {
+				if conversionError {
+					return
+				}
 
+				presenceTable, ok := v.(*lua.LTable)
+				if !ok {
+					conversionError = true
+					l.ArgError(3, "expects a valid set of presences")
+					return
+				}
+
+				presenceID := &PresenceID{}
+				presenceTable.ForEach(func(k lua.LValue, v lua.LValue) {
+					if conversionError {
+						return
+					}
+
+					switch k.String() {
+					case "session_id":
+						if v.Type() != lua.LTString {
+							conversionError = true
+							l.ArgError(3, "presence session id must be a string")
+							return
+						}
+						presenceID.SessionID, err = uuid.FromString(v.String())
+						if err != nil {
+							conversionError = true
+							l.ArgError(3, "presence session id must be a valid identifier")
+							return
+						}
+					case "node_id":
+						if v.Type() != lua.LTString {
+							conversionError = true
+							l.ArgError(3, "presence node id must be a string")
+							return
+						}
+						presenceID.Node = v.String()
+					}
+				})
+				if conversionError {
+					return
+				}
+
+				if presenceID.Node == "" {
+					presenceID.Node = n.node
+				}
+
+				presenceIDs = append(presenceIDs, presenceID)
+			})
+		}
+	}
+	if conversionError {
+		return 0
+	}
+
+	if len(presenceIDs) == 0 {
+		// Sending to whole stream.
+		n.router.SendToStream(n.logger, stream, msg)
+	} else {
+		// Sending to a subset of stream users.
+		n.router.SendToPresenceIDs(n.logger, presenceIDs, true, stream.Mode, msg)
+	}
+
+	return 0
+}
+
+func (n *RuntimeLuaNakamaModule) sessionDisconnect(l *lua.LState) int {
+	// Parse input Session ID.
+	sessionIDString := l.CheckString(1)
+	if sessionIDString == "" {
+		l.ArgError(1, "expects session id")
+		return 0
+	}
+	sessionID, err := uuid.FromString(sessionIDString)
+	if err != nil {
+		l.ArgError(1, "expects valid session id")
+		return 0
+	}
+
+	node := l.OptString(2, n.node)
+
+	if err := n.sessionRegistry.Disconnect(l.Context(), n.logger, sessionID, node); err != nil {
+		l.RaiseError(fmt.Sprintf("failed to disconnect: %s", err.Error()))
+	}
 	return 0
 }