+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ## [Unreleased]
 ### Added
 - Additional logging format option for Stackdriver Logging.
+- New runtime function to immediately disconnect active sockets.
+
+### Changed
+- Runtime stream send functions can now send to both the whole stream or only to a subset of users.
 
 ### Fixed
 - Correctly return group user results in Lua runtime listing operation.